+++ src/const.js
@@ -7,11 +7,6 @@ const MAX_COUNT_SENTENCES = 5;
 const MAX_COUNT_PHOTOS = 7;
 const MAX_PRICE_OFFER = 50;
 
-const RenderPosition = {
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-};
-
 const POINT_BLANK = {
   id: '',
   typePoint: '',
@@ -97,4 +92,4 @@ const OFFER_TITLES_TYPES = [
   },
 ];
 
-export {TRIP_POINTS_COUNT, RenderPosition, POINT_BLANK, POINT_TYPES, MAX_DAYS_GAP, MAX_MINUTES_GAP, POINT_CITIES, OFFER_TITLES_TYPES, TEXT_DESCRIPTION, URL_PHOTO, MAX_COUNT_SENTENCES, MAX_COUNT_PHOTOS, MAX_PRICE_OFFER};
+export {TRIP_POINTS_COUNT, POINT_BLANK, POINT_TYPES, MAX_DAYS_GAP, MAX_MINUTES_GAP, POINT_CITIES, OFFER_TITLES_TYPES, TEXT_DESCRIPTION, URL_PHOTO, MAX_COUNT_SENTENCES, MAX_COUNT_PHOTOS, MAX_PRICE_OFFER};
